Added Code for healthbar(not complete)
Implemented healthbar sprite and method of calculating how full the bar should be based on player health. Because we want max health to be variable, I've made it to where the healthbar represents the percent of health remaining, rounded down to the nearest percent cleanly divisible by 5. I ran into some issues displaying the healthbar, but I've thought of a solution that I'll be implementing in the next commit.